refactor: add remote folder to better organize experiment module
