+++ crates/tracel-core/src/experiment/remote/cloud/mod.rs
@@ -21,7 +21,7 @@ use tracel_experiment::{CancelToken, ExperimentId, ExperimentRun};
 
 use crate::backend::cloud::CloudBackend;
 use crate::experiment::ExperimentProvider;
-use crate::experiment::session::RemoteExperimentSession;
+use crate::experiment::remote::session::RemoteExperimentSession;
 
 #[derive(Debug, Clone)]
 pub(crate) struct ExperimentPath {
